Fix inconsistent typing in pbespgsolve, which is an issue on 32-bit Ubuntu.
Also added a check for the presence of unordered_set. If this is available
on all supported platforms, we can drop all the exceptions in compatibility.h
of the pbespgsolve tool.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12324 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/pbespgsolve/ParityGame_IO.cpp
@@ -201,8 +201,8 @@ void ParityGame::assign_pbes(mcrl2::pbes_system::pbes &pbes, verti *goal_vertex,
     verti begin = 0, end = 3;
     for (verti v = begin; v < end; ++v)
     {
-        std::set<long unsigned> deps = pgg.get_dependencies(v);
-        for ( std::set<long unsigned>::const_iterator it = deps.begin();
+        std::set<size_t> deps = pgg.get_dependencies(v);
+        for ( std::set<size_t>::const_iterator it = deps.begin();
                 it != deps.end(); ++it )
         {
             verti w = (verti)*it;
